CR fixes

+++ src/main/java/com/descope/model/inbound/InboundAppConsent.java
@@ -15,7 +15,7 @@ public class InboundAppConsent {
   private String userId;
   private String []scopes;
   private String grantedBy;
-  private int createdTime;
+  private long createdTime;
   private String tenantId;
 }
